refactor(settings): split monolithic SettingsDialog into modular tab components
- Create GeneralTab.tsx with app info and keyboard shortcuts
- Create AISettingsTab.tsx for OpenAI API key management
- Create GitHubSettingsTab.tsx for GitHub token configuration
- Create TroubleshootingTab.tsx for developer mode and logs
- Refactor SettingsDialog.tsx to shell component (~200 lines)
- Preserve tab state by hiding inactive tabs instead of unmounting
- Add loading/error guards matching AgentSettings pattern
- Add unmount safety flags for async operations
- Improve error handling with try/catch and console logging
