+++ src/components/Settings/SettingsDialog.tsx
@@ -2,37 +2,21 @@
  * Settings Dialog Component
  *
  * Modal UI for viewing and configuring application settings.
- * Includes tabs for General info, AI settings, and Troubleshooting tools.
+ * Contains a tabbed interface with navigation sidebar.
+ * Tab-specific content is rendered by dedicated tab components.
  */
 
-import { useState, useEffect, useCallback } from "react";
-import { Button } from "@/components/ui/button";
+import { useState, useEffect } from "react";
 import { useErrors } from "@/hooks";
 import { useLogsStore } from "@/store";
-import {
-  X,
-  FileText,
-  Trash2,
-  Key,
-  Check,
-  AlertCircle,
-  Loader2,
-  Sparkles,
-  FlaskConical,
-  TreePine,
-  Bot,
-  Bug,
-  ChevronDown,
-  ChevronRight,
-  Github,
-  ExternalLink,
-} from "lucide-react";
+import { X, Sparkles, Bot, Github } from "lucide-react";
 import { cn } from "@/lib/utils";
-import type { AIServiceState, GitHubTokenConfig } from "@/types";
-import { appClient, aiClient, logsClient, githubClient } from "@/clients";
+import { appClient } from "@/clients";
 import { AgentSettings } from "./AgentSettings";
-import { getStateDebugEnabled, enableStateDebug, disableStateDebug } from "@/components/Terminal";
-import type { AppState } from "@shared/types";
+import { GeneralTab } from "./GeneralTab";
+import { AISettingsTab } from "./AISettingsTab";
+import { GitHubSettingsTab } from "./GitHubSettingsTab";
+import { TroubleshootingTab } from "./TroubleshootingTab";
 
 interface SettingsDialogProps {
   isOpen: boolean;
@@ -43,67 +27,6 @@ interface SettingsDialogProps {
 }
 
 type SettingsTab = "general" | "agents" | "ai" | "github" | "troubleshooting";
-
-// Keyboard shortcuts organized by category
-const KEYBOARD_SHORTCUTS = [
-  {
-    category: "Agents",
-    shortcuts: [
-      { key: "Ctrl+Shift+C", description: "Start Claude agent" },
-      { key: "Ctrl+Shift+G", description: "Start Gemini agent" },
-      { key: "Ctrl+Shift+I", description: "Inject context to agent" },
-      { key: "Cmd+T", description: "Open agent palette" },
-    ],
-  },
-  {
-    category: "Navigation",
-    shortcuts: [
-      { key: "Ctrl+Tab", description: "Focus next agent or shell" },
-      { key: "Ctrl+Shift+Tab", description: "Focus previous agent or shell" },
-      { key: "Ctrl+Shift+F", description: "Toggle maximize focused tile" },
-    ],
-  },
-  {
-    category: "Panels",
-    shortcuts: [
-      { key: "Ctrl+Shift+L", description: "Toggle logs panel" },
-      { key: "Ctrl+Shift+E", description: "Toggle event inspector" },
-    ],
-  },
-  {
-    category: "Other",
-    shortcuts: [
-      { key: "Cmd+K Z", description: "Toggle focus mode (chord: press Cmd+K, release, then Z)" },
-    ],
-  },
-];
-
-const AI_MODELS = [
-  {
-    value: "gpt-5-nano",
-    label: "GPT-5 Nano",
-    description: "Fastest and most cost-effective (recommended)",
-  },
-  { value: "gpt-5-mini", label: "GPT-5 Mini", description: "Balanced speed and capability" },
-  { value: "gpt-5.1", label: "GPT-5.1", description: "Most capable flagship model" },
-];
-
-// Format key for platform-specific display
-const formatKey = (key: string): string => {
-  // Use process.platform from Electron for reliable platform detection
-  const isMac = window.navigator.platform.toUpperCase().indexOf("MAC") >= 0;
-
-  if (isMac) {
-    return key
-      .replace(/Cmd\+/g, "⌘")
-      .replace(/Ctrl\+/g, "⌃")
-      .replace(/Shift\+/g, "⇧")
-      .replace(/Alt\+/g, "⌥");
-  }
-
-  // On Windows/Linux, replace Cmd with Ctrl
-  return key.replace(/Cmd\+/g, "Ctrl+");
-};
 
 export function SettingsDialog({
   isOpen,
@@ -118,163 +41,6 @@ export function SettingsDialog({
   // App version state
   const [appVersion, setAppVersion] = useState<string>("Loading...");
 
-  // AI settings state
-  const [aiConfig, setAiConfig] = useState<AIServiceState | null>(null);
-  const [apiKey, setApiKey] = useState("");
-  const [isValidating, setIsValidating] = useState(false);
-  const [isTesting, setIsTesting] = useState(false);
-  const [validationResult, setValidationResult] = useState<
-    "success" | "error" | "test-success" | "test-error" | null
-  >(null);
-  const [selectedModel, setSelectedModel] = useState("gpt-5-nano");
-
-  // GitHub settings state
-  const [githubConfig, setGithubConfig] = useState<GitHubTokenConfig | null>(null);
-  const [githubToken, setGithubToken] = useState("");
-  const [isGithubValidating, setIsGithubValidating] = useState(false);
-  const [isGithubTesting, setIsGithubTesting] = useState(false);
-  const [githubValidationResult, setGithubValidationResult] = useState<
-    "success" | "error" | "test-success" | "test-error" | null
-  >(null);
-  const [githubErrorMessage, setGithubErrorMessage] = useState<string | null>(null);
-
-  // Developer mode settings
-  const [developerMode, setDeveloperMode] = useState(false);
-  const [showStateDebug, setShowStateDebug] = useState(() => getStateDebugEnabled());
-  const [autoOpenDiagnostics, setAutoOpenDiagnostics] = useState(false);
-  const [focusEventsTab, setFocusEventsTab] = useState(false);
-
-  // Collapsible section states
-  const [isAdvancedOpen, setIsAdvancedOpen] = useState(false);
-  const [isShortcutsOpen, setIsShortcutsOpen] = useState(false);
-
-  // Save developer mode settings to app state
-  const saveDeveloperModeSettings = useCallback(
-    async (settings: NonNullable<AppState["developerMode"]>) => {
-      try {
-        await appClient.setState({ developerMode: settings });
-      } catch (error) {
-        console.error("Failed to save developer mode settings:", error);
-      }
-    },
-    []
-  );
-
-  // Handle master developer mode toggle
-  const handleToggleDeveloperMode = useCallback(() => {
-    const newEnabled = !developerMode;
-    setDeveloperMode(newEnabled);
-
-    // If disabling, turn off all child features
-    if (!newEnabled) {
-      setShowStateDebug(false);
-      disableStateDebug();
-      // Dispatch event to notify other components that debug mode is off
-      if (typeof window !== "undefined") {
-        window.dispatchEvent(
-          new CustomEvent("canopy:debug-toggle", { detail: { enabled: false } })
-        );
-      }
-      setAutoOpenDiagnostics(false);
-      setFocusEventsTab(false);
-      saveDeveloperModeSettings({
-        enabled: false,
-        showStateDebug: false,
-        autoOpenDiagnostics: false,
-        focusEventsTab: false,
-      });
-    } else {
-      saveDeveloperModeSettings({
-        enabled: true,
-        showStateDebug,
-        autoOpenDiagnostics,
-        focusEventsTab,
-      });
-    }
-  }, [
-    developerMode,
-    showStateDebug,
-    autoOpenDiagnostics,
-    focusEventsTab,
-    saveDeveloperModeSettings,
-  ]);
-
-  // Handle state debug toggle
-  const handleToggleStateDebug = useCallback(() => {
-    const newState = !showStateDebug;
-    setShowStateDebug(newState);
-    if (newState) {
-      enableStateDebug();
-    } else {
-      disableStateDebug();
-    }
-    // Dispatch event to notify other components (e.g., DebugInfo)
-    if (typeof window !== "undefined") {
-      window.dispatchEvent(
-        new CustomEvent("canopy:debug-toggle", { detail: { enabled: newState } })
-      );
-    }
-    saveDeveloperModeSettings({
-      enabled: developerMode,
-      showStateDebug: newState,
-      autoOpenDiagnostics,
-      focusEventsTab,
-    });
-  }, [
-    showStateDebug,
-    developerMode,
-    autoOpenDiagnostics,
-    focusEventsTab,
-    saveDeveloperModeSettings,
-  ]);
-
-  // Handle auto-open diagnostics toggle
-  const handleToggleAutoOpenDiagnostics = useCallback(() => {
-    const newState = !autoOpenDiagnostics;
-    setAutoOpenDiagnostics(newState);
-    // If disabling auto-open, also disable focus events tab
-    if (!newState) {
-      setFocusEventsTab(false);
-      saveDeveloperModeSettings({
-        enabled: developerMode,
-        showStateDebug,
-        autoOpenDiagnostics: false,
-        focusEventsTab: false,
-      });
-    } else {
-      saveDeveloperModeSettings({
-        enabled: developerMode,
-        showStateDebug,
-        autoOpenDiagnostics: true,
-        focusEventsTab,
-      });
-    }
-  }, [
-    autoOpenDiagnostics,
-    developerMode,
-    showStateDebug,
-    focusEventsTab,
-    saveDeveloperModeSettings,
-  ]);
-
-  // Handle focus events tab toggle
-  const handleToggleFocusEventsTab = useCallback(() => {
-    const newState = !focusEventsTab;
-    setFocusEventsTab(newState);
-    saveDeveloperModeSettings({
-      enabled: developerMode,
-      showStateDebug,
-      autoOpenDiagnostics,
-      focusEventsTab: newState,
-    });
-  }, [
-    focusEventsTab,
-    developerMode,
-    showStateDebug,
-    autoOpenDiagnostics,
-    saveDeveloperModeSettings,
-  ]);
-
   // Update active tab when defaultTab changes while dialog is open
   useEffect(() => {
     if (isOpen && defaultTab && defaultTab !== activeTab) {
@@ -282,14 +48,6 @@ export function SettingsDialog({
     }
     // eslint-disable-next-line react-hooks/exhaustive-deps
   }, [isOpen, defaultTab]);
-
-  // Reset collapsible sections when dialog closes
-  useEffect(() => {
-    if (!isOpen) {
-      setIsAdvancedOpen(false);
-      setIsShortcutsOpen(false);
-    }
-  }, [isOpen]);
 
   // Load app version on mount
   useEffect(() => {
@@ -303,189 +61,6 @@ export function SettingsDialog({
         });
     }
   }, [isOpen]);
-
-  // Load AI config and developer mode settings on mount
-  useEffect(() => {
-    if (!isOpen) return;
-
-    // Load AI config
-    aiClient.getConfig().then((config) => {
-      setAiConfig(config);
-      setSelectedModel(config.model);
-    });
-
-    // Load developer mode settings
-    appClient.getState().then((appState) => {
-      if (appState?.developerMode) {
-        // Use persisted developer mode settings
-        setDeveloperMode(appState.developerMode.enabled);
-        setShowStateDebug(appState.developerMode.showStateDebug);
-        setAutoOpenDiagnostics(appState.developerMode.autoOpenDiagnostics);
-        setFocusEventsTab(appState.developerMode.focusEventsTab);
-        // Sync localStorage with persisted state
-        if (appState.developerMode.showStateDebug) {
-          enableStateDebug();
-        } else {
-          disableStateDebug();
-        }
-      } else {
-        // No persisted settings - sync from localStorage
-        setShowStateDebug(getStateDebugEnabled());
-      }
-    });
-  }, [isOpen]);
-
-  // Load GitHub config on mount
-  useEffect(() => {
-    if (isOpen) {
-      githubClient.getConfig().then(setGithubConfig);
-    }
-  }, [isOpen]);
-
-  // Clear validation result after 3 seconds
-  useEffect(() => {
-    if (!validationResult) return;
-    const timer = setTimeout(() => setValidationResult(null), 3000);
-    return () => clearTimeout(timer);
-  }, [validationResult]);
-
-  // Clear GitHub validation result after 3 seconds
-  useEffect(() => {
-    if (!githubValidationResult) return;
-    const timer = setTimeout(() => {
-      setGithubValidationResult(null);
-      setGithubErrorMessage(null);
-    }, 5000);
-    return () => clearTimeout(timer);
-  }, [githubValidationResult]);
-
-  const handleClearLogs = async () => {
-    clearLogs();
-    await logsClient.clear();
-  };
-
-  const handleSaveKey = async () => {
-    if (!apiKey.trim()) return;
-
-    setIsValidating(true);
-    setValidationResult(null);
-
-    try {
-      const success = await aiClient.setKey(apiKey.trim());
-      if (success) {
-        setApiKey(""); // Clear input for security
-        setValidationResult("success");
-        // Refresh config
-        const config = await aiClient.getConfig();
-        setAiConfig(config);
-      } else {
-        setValidationResult("error");
-      }
-    } catch {
-      setValidationResult("error");
-    } finally {
-      setIsValidating(false);
-    }
-  };
-
-  const handleClearKey = async () => {
-    await aiClient.clearKey();
-    const config = await aiClient.getConfig();
-    setAiConfig(config);
-    setValidationResult(null);
-  };
-
-  const handleTestKey = async () => {
-    if (!apiKey.trim()) return;
-
-    setIsTesting(true);
-    setValidationResult(null);
-
-    try {
-      const isValid = await aiClient.validateKey(apiKey.trim());
-      setValidationResult(isValid ? "test-success" : "test-error");
-    } catch {
-      setValidationResult("test-error");
-    } finally {
-      setIsTesting(false);
-    }
-  };
-
-  const handleModelChange = async (model: string) => {
-    setSelectedModel(model);
-    await aiClient.setModel(model);
-    const config = await aiClient.getConfig();
-    setAiConfig(config);
-  };
-
-  const handleEnabledChange = async (enabled: boolean) => {
-    await aiClient.setEnabled(enabled);
-    const config = await aiClient.getConfig();
-    setAiConfig(config);
-  };
-
-  // GitHub token handlers
-  const handleGitHubSaveToken = async () => {
-    if (!githubToken.trim()) return;
-
-    setIsGithubValidating(true);
-    setGithubValidationResult(null);
-    setGithubErrorMessage(null);
-
-    try {
-      const result = await githubClient.setToken(githubToken.trim());
-      if (result.valid) {
-        setGithubToken(""); // Clear input for security
-        setGithubValidationResult("success");
-        // Refresh config
-        const config = await githubClient.getConfig();
-        setGithubConfig(config);
-      } else {
-        setGithubValidationResult("error");
-        setGithubErrorMessage(result.error || "Invalid token");
-      }
-    } catch {
-      setGithubValidationResult("error");
-      setGithubErrorMessage("Failed to save token");
-    } finally {
-      setIsGithubValidating(false);
-    }
-  };
-
-  const handleGitHubClearToken = async () => {
-    await githubClient.clearToken();
-    const config = await githubClient.getConfig();
-    setGithubConfig(config);
-    setGithubValidationResult(null);
-    setGithubErrorMessage(null);
-  };
-
-  const handleGitHubTestToken = async () => {
-    if (!githubToken.trim()) return;
-
-    setIsGithubTesting(true);
-    setGithubValidationResult(null);
-    setGithubErrorMessage(null);
-
-    try {
-      const result = await githubClient.validateToken(githubToken.trim());
-      setGithubValidationResult(result.valid ? "test-success" : "test-error");
-      if (!result.valid) {
-        setGithubErrorMessage(result.error || "Invalid token");
-      }
-    } catch {
-      setGithubValidationResult("test-error");
-      setGithubErrorMessage("Failed to validate token");
-    } finally {
-      setIsGithubTesting(false);
-    }
-  };
-
-  const openGitHubTokenPage = () => {
-    window.electron.system.openExternal(
-      "https://github.com/settings/tokens/new?scopes=repo,read:org&description=Canopy%20Command%20Center"
-    );
-  };
 
   if (!isOpen) return null;
 
@@ -588,588 +163,26 @@ export function SettingsDialog({
           </div>
 
           <div className="p-6 overflow-y-auto flex-1">
-            {activeTab === "general" && (
-              <div className="space-y-6">
-                <div className="space-y-2">
-                  <h4 className="text-sm font-medium text-canopy-text">About</h4>
-                  <div className="bg-canopy-bg border border-canopy-border rounded-md p-4">
-                    <div className="flex items-center gap-3 mb-4">
-                      <div className="h-12 w-12 bg-canopy-accent/20 rounded-lg flex items-center justify-center">
-                        <TreePine className="w-6 h-6 text-canopy-accent" />
-                      </div>
-                      <div>
-                        <div className="font-semibold text-canopy-text text-lg">Canopy</div>
-                        <div className="text-sm text-gray-400">Command Center</div>
-                      </div>
-                    </div>
-                    <div className="space-y-2 text-sm">
-                      <div className="flex justify-between text-gray-400">
-                        <span>Version</span>
-                        <span className="font-mono text-canopy-text">{appVersion}</span>
-                      </div>
-                    </div>
-                  </div>
-                </div>
+            {/* Render all tabs but hide inactive ones to preserve state */}
+            <div className={activeTab === "general" ? "" : "hidden"}>
+              <GeneralTab appVersion={appVersion} />
+            </div>
 
-                <div className="space-y-2">
-                  <h4 className="text-sm font-medium text-canopy-text">Description</h4>
-                  <p className="text-sm text-gray-400">
-                    An orchestration board for AI coding agents. Start agents on worktrees, monitor
-                    their progress, and inject context to help them understand your codebase.
-                  </p>
-                </div>
+            <div className={activeTab === "agents" ? "" : "hidden"}>
+              <AgentSettings onSettingsChange={onSettingsChange} />
+            </div>
 
-                {/* Keyboard Shortcuts (collapsed by default) */}
-                <div className="border border-canopy-border rounded-md">
-                  <button
-                    type="button"
-                    onClick={() => setIsShortcutsOpen(!isShortcutsOpen)}
-                    aria-expanded={isShortcutsOpen}
-                    aria-controls="keyboard-shortcuts-content"
-                    className="w-full flex items-center gap-2 px-3 py-2 text-sm text-gray-400 hover:text-canopy-text transition-colors"
-                  >
-                    {isShortcutsOpen ? (
-                      <ChevronDown className="w-4 h-4" />
-                    ) : (
-                      <ChevronRight className="w-4 h-4" />
-                    )}
-                    <span>Keyboard Shortcuts</span>
-                  </button>
+            <div className={activeTab === "ai" ? "" : "hidden"}>
+              <AISettingsTab />
+            </div>
 
-                  {isShortcutsOpen && (
-                    <div
-                      id="keyboard-shortcuts-content"
-                      className="px-3 pb-3 space-y-4 border-t border-canopy-border pt-3"
-                    >
-                      {KEYBOARD_SHORTCUTS.map((category) => (
-                        <div key={category.category} className="space-y-2">
-                          <h5 className="text-xs font-medium text-gray-400 uppercase tracking-wide">
-                            {category.category}
-                          </h5>
-                          <dl className="space-y-1">
-                            {category.shortcuts.map((shortcut) => (
-                              <div
-                                key={shortcut.key}
-                                className="flex items-center justify-between text-sm py-1"
-                              >
-                                <dt className="text-gray-300">{shortcut.description}</dt>
-                                <dd>
-                                  <kbd className="px-2 py-1 bg-canopy-bg border border-canopy-border rounded text-xs font-mono text-canopy-text">
-                                    {formatKey(shortcut.key)}
-                                  </kbd>
-                                </dd>
-                              </div>
-                            ))}
-                          </dl>
-                        </div>
-                      ))}
-                    </div>
-                  )}
-                </div>
-              </div>
-            )}
+            <div className={activeTab === "github" ? "" : "hidden"}>
+              <GitHubSettingsTab />
+            </div>
 
-            {activeTab === "agents" && <AgentSettings onSettingsChange={onSettingsChange} />}
-
-            {activeTab === "ai" && (
-              <div className="space-y-6">
-                {/* API Key Section */}
-                <div className="space-y-3">
-                  <div className="flex items-center justify-between">
-                    <h4 className="text-sm font-medium text-canopy-text flex items-center gap-2">
-                      <Key className="w-4 h-4" />
-                      OpenAI API Key
-                    </h4>
-                    {aiConfig?.hasKey && (
-                      <span className="text-xs text-[var(--color-status-success)] flex items-center gap-1">
-                        <Check className="w-3 h-3" />
-                        Key configured
-                      </span>
-                    )}
-                  </div>
-
-                  <div className="flex gap-2">
-                    <input
-                      type="password"
-                      value={apiKey}
-                      onChange={(e) => setApiKey(e.target.value)}
-                      placeholder={aiConfig?.hasKey ? "Enter new key to replace" : "sk-..."}
-                      className="flex-1 bg-canopy-bg border border-canopy-border rounded-md px-3 py-2 text-sm text-canopy-text placeholder:text-gray-500 focus:outline-none focus:ring-1 focus:ring-canopy-accent"
-                      disabled={isValidating || isTesting}
-                    />
-                    <Button
-                      onClick={handleTestKey}
-                      disabled={isTesting || isValidating || !apiKey.trim()}
-                      variant="outline"
-                      size="sm"
-                      className="min-w-[70px] text-canopy-text border-canopy-border hover:bg-canopy-border"
-                    >
-                      {isTesting ? (
-                        <Loader2 className="w-4 h-4 animate-spin" />
-                      ) : (
-                        <>
-                          <FlaskConical className="w-4 h-4 mr-1" />
-                          Test
-                        </>
-                      )}
-                    </Button>
-                    <Button
-                      onClick={handleSaveKey}
-                      disabled={isValidating || isTesting || !apiKey.trim()}
-                      size="sm"
-                      className="min-w-[70px]"
-                    >
-                      {isValidating ? <Loader2 className="w-4 h-4 animate-spin" /> : "Save"}
-                    </Button>
-                    {aiConfig?.hasKey && (
-                      <Button
-                        onClick={handleClearKey}
-                        variant="outline"
-                        size="sm"
-                        className="text-[var(--color-status-error)] border-canopy-border hover:bg-red-900/20 hover:text-red-300 hover:border-red-900/30"
-                      >
-                        Clear
-                      </Button>
-                    )}
-                  </div>
-
-                  {validationResult === "success" && (
-                    <p className="text-xs text-[var(--color-status-success)] flex items-center gap-1">
-                      <Check className="w-3 h-3" />
-                      API key validated and saved successfully
-                    </p>
-                  )}
-                  {validationResult === "test-success" && (
-                    <p className="text-xs text-[var(--color-status-success)] flex items-center gap-1">
-                      <Check className="w-3 h-3" />
-                      API key is valid! Click Save to store it.
-                    </p>
-                  )}
-                  {validationResult === "error" && (
-                    <p className="text-xs text-[var(--color-status-error)] flex items-center gap-1">
-                      <AlertCircle className="w-3 h-3" />
-                      Invalid API key. Please check and try again.
-                    </p>
-                  )}
-                  {validationResult === "test-error" && (
-                    <p className="text-xs text-[var(--color-status-error)] flex items-center gap-1">
-                      <AlertCircle className="w-3 h-3" />
-                      API key test failed. Please check your key.
-                    </p>
-                  )}
-
-                  <p className="text-xs text-gray-500">
-                    Used for worktree summaries and project identity. Helps agents understand your
-                    codebase context. Stored locally and never sent to our servers.
-                  </p>
-                </div>
-
-                {/* Advanced Options (collapsed by default) */}
-                <div className="border border-canopy-border rounded-md">
-                  <button
-                    type="button"
-                    onClick={() => setIsAdvancedOpen(!isAdvancedOpen)}
-                    aria-expanded={isAdvancedOpen}
-                    aria-controls="advanced-options-content"
-                    className="w-full flex items-center gap-2 px-3 py-2 text-sm text-gray-400 hover:text-canopy-text transition-colors"
-                  >
-                    {isAdvancedOpen ? (
-                      <ChevronDown className="w-4 h-4" />
-                    ) : (
-                      <ChevronRight className="w-4 h-4" />
-                    )}
-                    <span>Advanced Options</span>
-                  </button>
-
-                  {isAdvancedOpen && (
-                    <div
-                      id="advanced-options-content"
-                      className="px-3 pb-3 space-y-4 border-t border-canopy-border pt-3"
-                    >
-                      {/* Model Selection */}
-                      <div className="space-y-3">
-                        <h4 className="text-sm font-medium text-canopy-text">AI Model</h4>
-                        <div className="space-y-2">
-                          {AI_MODELS.map((model) => (
-                            <label
-                              key={model.value}
-                              className={cn(
-                                "flex items-center gap-3 p-3 rounded-md border cursor-pointer transition-colors",
-                                selectedModel === model.value
-                                  ? "border-canopy-accent bg-canopy-accent/10"
-                                  : "border-canopy-border hover:border-gray-500"
-                              )}
-                            >
-                              <input
-                                type="radio"
-                                name="ai-model"
-                                value={model.value}
-                                checked={selectedModel === model.value}
-                                onChange={() => handleModelChange(model.value)}
-                                className="sr-only"
-                              />
-                              <div
-                                className={cn(
-                                  "w-4 h-4 rounded-full border-2 flex items-center justify-center",
-                                  selectedModel === model.value
-                                    ? "border-canopy-accent"
-                                    : "border-gray-500"
-                                )}
-                              >
-                                {selectedModel === model.value && (
-                                  <div className="w-2 h-2 rounded-full bg-canopy-accent" />
-                                )}
-                              </div>
-                              <div className="flex-1">
-                                <div className="text-sm font-medium text-canopy-text">
-                                  {model.label}
-                                </div>
-                                <div className="text-xs text-gray-500">{model.description}</div>
-                              </div>
-                            </label>
-                          ))}
-                        </div>
-                        <p className="text-xs text-gray-500">
-                          GPT-5 Nano is recommended for most tasks.
-                        </p>
-                      </div>
-
-                      {/* Enable/Disable Toggle */}
-                      <div className="space-y-3">
-                        <h4 className="text-sm font-medium text-canopy-text">AI Features</h4>
-                        <label className="flex items-center gap-3 cursor-pointer">
-                          <button
-                            onClick={() => handleEnabledChange(!aiConfig?.enabled)}
-                            className={cn(
-                              "relative w-11 h-6 rounded-full transition-colors",
-                              aiConfig?.enabled ? "bg-canopy-accent" : "bg-gray-600"
-                            )}
-                          >
-                            <span
-                              className={cn(
-                                "absolute top-0.5 left-0.5 w-5 h-5 bg-white rounded-full transition-transform",
-                                aiConfig?.enabled && "translate-x-5"
-                              )}
-                            />
-                          </button>
-                          <span className="text-sm text-canopy-text">
-                            {aiConfig?.enabled ? "Enabled" : "Disabled"}
-                          </span>
-                        </label>
-                        <p className="text-xs text-gray-500">
-                          When enabled, Canopy generates worktree summaries and project identities
-                          to help agents understand your work.
-                        </p>
-                      </div>
-                    </div>
-                  )}
-                </div>
-              </div>
-            )}
-
-            {activeTab === "github" && (
-              <div className="space-y-6">
-                {/* Token Section */}
-                <div className="space-y-3">
-                  <div className="flex items-center justify-between">
-                    <h4 className="text-sm font-medium text-canopy-text flex items-center gap-2">
-                      <Key className="w-4 h-4" />
-                      Personal Access Token
-                    </h4>
-                    {githubConfig?.hasToken && (
-                      <span className="text-xs text-[var(--color-status-success)] flex items-center gap-1">
-                        <Check className="w-3 h-3" />
-                        Token configured
-                      </span>
-                    )}
-                  </div>
-
-                  <div className="flex gap-2">
-                    <input
-                      type="password"
-                      value={githubToken}
-                      onChange={(e) => setGithubToken(e.target.value)}
-                      placeholder={
-                        githubConfig?.hasToken
-                          ? "Enter new token to replace"
-                          : "ghp_... or github_pat_..."
-                      }
-                      className="flex-1 bg-canopy-bg border border-canopy-border rounded-md px-3 py-2 text-sm text-canopy-text placeholder:text-gray-500 focus:outline-none focus:ring-1 focus:ring-canopy-accent"
-                      disabled={isGithubValidating || isGithubTesting}
-                    />
-                    <Button
-                      onClick={handleGitHubTestToken}
-                      disabled={isGithubTesting || isGithubValidating || !githubToken.trim()}
-                      variant="outline"
-                      size="sm"
-                      className="min-w-[70px] text-canopy-text border-canopy-border hover:bg-canopy-border"
-                    >
-                      {isGithubTesting ? (
-                        <Loader2 className="w-4 h-4 animate-spin" />
-                      ) : (
-                        <>
-                          <FlaskConical className="w-4 h-4 mr-1" />
-                          Test
-                        </>
-                      )}
-                    </Button>
-                    <Button
-                      onClick={handleGitHubSaveToken}
-                      disabled={isGithubValidating || isGithubTesting || !githubToken.trim()}
-                      size="sm"
-                      className="min-w-[70px]"
-                    >
-                      {isGithubValidating ? <Loader2 className="w-4 h-4 animate-spin" /> : "Save"}
-                    </Button>
-                    {githubConfig?.hasToken && (
-                      <Button
-                        onClick={handleGitHubClearToken}
-                        variant="outline"
-                        size="sm"
-                        className="text-[var(--color-status-error)] border-canopy-border hover:bg-red-900/20 hover:text-red-300 hover:border-red-900/30"
-                      >
-                        Clear
-                      </Button>
-                    )}
-                  </div>
-
-                  {githubValidationResult === "success" && (
-                    <p className="text-xs text-[var(--color-status-success)] flex items-center gap-1">
-                      <Check className="w-3 h-3" />
-                      Token validated and saved successfully
-                    </p>
-                  )}
-                  {githubValidationResult === "test-success" && (
-                    <p className="text-xs text-[var(--color-status-success)] flex items-center gap-1">
-                      <Check className="w-3 h-3" />
-                      Token is valid! Click Save to store it.
-                    </p>
-                  )}
-                  {githubValidationResult === "error" && (
-                    <p className="text-xs text-[var(--color-status-error)] flex items-center gap-1">
-                      <AlertCircle className="w-3 h-3" />
-                      {githubErrorMessage || "Invalid token. Please check and try again."}
-                    </p>
-                  )}
-                  {githubValidationResult === "test-error" && (
-                    <p className="text-xs text-[var(--color-status-error)] flex items-center gap-1">
-                      <AlertCircle className="w-3 h-3" />
-                      {githubErrorMessage || "Token test failed. Please check your token."}
-                    </p>
-                  )}
-
-                  <p className="text-xs text-gray-500">
-                    Used for repository statistics, issue/PR detection, and linking worktrees to
-                    GitHub. Eliminates the need for the gh CLI.
-                  </p>
-                </div>
-
-                {/* Create Token Section */}
-                <div className="space-y-3 border border-canopy-border rounded-md p-4">
-                  <h4 className="text-sm font-medium text-canopy-text">Create a New Token</h4>
-                  <p className="text-xs text-gray-400">
-                    To create a personal access token with the required scopes (repo, read:org),
-                    click the button below. This will open GitHub in your browser.
-                  </p>
-                  <Button
-                    onClick={openGitHubTokenPage}
-                    variant="outline"
-                    size="sm"
-                    className="text-canopy-text border-canopy-border hover:bg-canopy-border"
-                  >
-                    <ExternalLink className="w-4 h-4 mr-2" />
-                    Create Token on GitHub
-                  </Button>
-                  <div className="mt-2 space-y-1">
-                    <p className="text-xs text-gray-500">Required scopes:</p>
-                    <ul className="text-xs text-gray-500 list-disc list-inside">
-                      <li>
-                        <code className="text-canopy-text bg-canopy-bg px-1 rounded">repo</code> -
-                        Access repository data
-                      </li>
-                      <li>
-                        <code className="text-canopy-text bg-canopy-bg px-1 rounded">read:org</code>{" "}
-                        - Read organization membership (for private repos)
-                      </li>
-                    </ul>
-                  </div>
-                </div>
-              </div>
-            )}
-
-            {activeTab === "troubleshooting" && (
-              <div className="space-y-6">
-                {/* Application Logs Section */}
-                <div className="space-y-4">
-                  <div>
-                    <h4 className="text-sm font-medium text-canopy-text mb-1">Application Logs</h4>
-                    <p className="text-xs text-gray-400 mb-3">
-                      View internal application logs for debugging purposes.
-                    </p>
-                    <div className="flex gap-3">
-                      <Button
-                        variant="outline"
-                        size="sm"
-                        onClick={() => openLogs()}
-                        className="text-canopy-text border-canopy-border hover:bg-canopy-border hover:text-canopy-text"
-                      >
-                        <FileText className="w-4 h-4 mr-2" />
-                        Open Log File
-                      </Button>
-                      <Button
-                        variant="outline"
-                        size="sm"
-                        onClick={handleClearLogs}
-                        className="text-[var(--color-status-error)] border-canopy-border hover:bg-red-900/20 hover:text-red-300 hover:border-red-900/30"
-                      >
-                        <Trash2 className="w-4 h-4 mr-2" />
-                        Clear Logs
-                      </Button>
-                    </div>
-                  </div>
-                </div>
-
-                {/* Developer Mode Section */}
-                <div className="space-y-4">
-                  <div>
-                    <h4 className="text-sm font-medium text-canopy-text mb-1 flex items-center gap-2">
-                      <Bug className="w-4 h-4" />
-                      Developer Mode
-                    </h4>
-                    <p className="text-xs text-gray-400 mb-3">
-                      Enable enhanced debugging features for development and troubleshooting.
-                    </p>
-
-                    {/* Master Developer Mode Toggle */}
-                    <label className="flex items-center gap-3 cursor-pointer mb-4 p-3 border border-canopy-border rounded-md">
-                      <button
-                        onClick={handleToggleDeveloperMode}
-                        className={cn(
-                          "relative w-11 h-6 rounded-full transition-colors shrink-0",
-                          developerMode ? "bg-canopy-accent" : "bg-gray-600"
-                        )}
-                      >
-                        <span
-                          className={cn(
-                            "absolute top-0.5 left-0.5 w-5 h-5 bg-white rounded-full transition-transform",
-                            developerMode && "translate-x-5"
-                          )}
-                        />
-                      </button>
-                      <div>
-                        <span className="text-sm text-canopy-text font-medium">
-                          Enable Developer Mode
-                        </span>
-                        <p className="text-xs text-gray-400">
-                          Activates all debugging features below
-                        </p>
-                      </div>
-                    </label>
-
-                    {/* Individual Debug Features */}
-                    <div
-                      className={cn(
-                        "ml-4 space-y-3 border-l-2 border-canopy-border pl-4 transition-opacity",
-                        !developerMode && "opacity-50"
-                      )}
-                    >
-                      {/* State Debug Overlays */}
-                      <label className="flex items-center gap-3 cursor-pointer">
-                        <input
-                          type="checkbox"
-                          checked={showStateDebug}
-                          onChange={handleToggleStateDebug}
-                          disabled={!developerMode}
-                          className="w-4 h-4 rounded border-gray-600 bg-gray-700 text-canopy-accent focus:ring-canopy-accent focus:ring-offset-0 disabled:opacity-50"
-                        />
-                        <div>
-                          <span className="text-sm text-canopy-text">State Debug Overlays</span>
-                          <p className="text-xs text-gray-400">
-                            Show trigger source and confidence in terminal headers
-                          </p>
-                        </div>
-                      </label>
-
-                      {/* Auto-open Diagnostics */}
-                      <label className="flex items-center gap-3 cursor-pointer">
-                        <input
-                          type="checkbox"
-                          checked={autoOpenDiagnostics}
-                          onChange={handleToggleAutoOpenDiagnostics}
-                          disabled={!developerMode}
-                          className="w-4 h-4 rounded border-gray-600 bg-gray-700 text-canopy-accent focus:ring-canopy-accent focus:ring-offset-0 disabled:opacity-50"
-                        />
-                        <div>
-                          <span className="text-sm text-canopy-text">
-                            Auto-Open Diagnostics Dock
-                          </span>
-                          <p className="text-xs text-gray-400">
-                            Automatically open diagnostics panel on app startup
-                          </p>
-                        </div>
-                      </label>
-
-                      {/* Focus Events Tab */}
-                      <label
-                        className={cn(
-                          "flex items-center gap-3 cursor-pointer ml-4",
-                          !autoOpenDiagnostics && "opacity-50"
-                        )}
-                      >
-                        <input
-                          type="checkbox"
-                          checked={focusEventsTab}
-                          onChange={handleToggleFocusEventsTab}
-                          disabled={!developerMode || !autoOpenDiagnostics}
-                          className="w-4 h-4 rounded border-gray-600 bg-gray-700 text-canopy-accent focus:ring-canopy-accent focus:ring-offset-0 disabled:opacity-50"
-                        />
-                        <div>
-                          <span className="text-sm text-canopy-text">Focus Events Tab</span>
-                          <p className="text-xs text-gray-400">
-                            Default to Events tab when diagnostics opens
-                          </p>
-                        </div>
-                      </label>
-                    </div>
-
-                    {/* Environment Variable Hints */}
-                    <div className="mt-4 p-3 bg-canopy-border/30 rounded-md">
-                      <h5 className="text-xs font-medium text-canopy-text mb-2">
-                        Advanced: Main Process Logging
-                      </h5>
-                      <p className="text-xs text-gray-400 mb-2">
-                        For verbose main process logs, restart the app with environment variables:
-                      </p>
-                      <code className="block text-xs bg-canopy-bg p-2 rounded border border-canopy-border font-mono text-canopy-text">
-                        CANOPY_DEBUG=1 CANOPY_VERBOSE=1 npm run dev
-                      </code>
-                      <div className="mt-2 space-y-1">
-                        <p className="text-xs text-gray-400">
-                          <span className="font-medium text-canopy-text">CANOPY_DEBUG</span> —
-                          General logger verbosity
-                        </p>
-                        <p className="text-xs text-gray-400">
-                          <span className="font-medium text-canopy-text">CANOPY_VERBOSE</span> —
-                          Service-level debug output
-                        </p>
-                      </div>
-                    </div>
-                  </div>
-                </div>
-
-                {/* Keyboard Shortcuts Section */}
-                <div className="space-y-4">
-                  <div>
-                    <h4 className="text-sm font-medium text-canopy-text mb-1">
-                      Keyboard Shortcuts
-                    </h4>
-                    <p className="text-xs text-gray-400 mb-3">
-                      Use Cmd+Option+I (Mac) or Ctrl+Shift+I (Windows/Linux) to open DevTools.
-                    </p>
-                  </div>
-                </div>
-              </div>
-            )}
+            <div className={activeTab === "troubleshooting" ? "" : "hidden"}>
+              <TroubleshootingTab openLogs={openLogs} clearLogs={clearLogs} />
+            </div>
           </div>
         </div>
       </div>
